requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy>=2.0.0
httpx
python-dotenv
psycopg2-binary
